KK KKP KPP part 3 Fix : quiet search       a lot of bug Modify : combine 3 project

+++ Minishogi/FileMapping.h
@@ -3,7 +3,7 @@
 
 class FileMapping {
 private:
-	const size_t BUF_SIZE = (1 << (sizeof(short) * 8)) - 1;
+	const DWORD BUF_SIZE = (1 << (sizeof(short) * 8)) - 1;
 	HANDLE hMapFile = NULL;
 	LPVOID pBuf = NULL;
